+++ TinhDaHinh/BT2/src/com/company/MiengDatHinhThang.java
@@ -53,7 +53,7 @@ public class MiengDatHinhThang extends MiengDat{
     }
     @Override
     public double tinhTien(){
-        return this.donGia*this.tinhDienTich()*0.9;
+        return this.donGia*(long) this.tinhDienTich()*0.9;
     }
